+++ ice-20201109/core/Properties/AssemblyInfo.cs
@@ -20,7 +20,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("bd530200-2430-11f1-8ffd-ebc25bcb3563")]
+[assembly: Guid("e7c09070-2908-11f1-8aa3-c1b2d0aa9358")]
 
 // Version information for an assembly consists of the following four values:
 //
@@ -29,5 +29,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("6.13.0.0")]
-[assembly: AssemblyFileVersion("6.13.1.0")]
+[assembly: AssemblyVersion("6.13.1.0")]
+[assembly: AssemblyFileVersion("6.14.0.0")]